refactor(views): dedupe Swap/Foundation trio into shared utils/components
confirmPriceImpactWithoutFee, SwapRoute, and ProgressSteps were
byte-identical between views/Swap/components/ and views/Foundation/
Donate/components/ (only the relative depth of config/constants
imports differed). Move all three to shared locations:

  * utils/confirmPriceImpactWithoutFee.ts
  * components/SwapRoute.tsx
  * components/ProgressSteps.tsx

The four consumer files (views/Swap/index.tsx,
views/Foundation/Donate/index.tsx, and both views/*/components/
AdvancedSwapDetails.tsx) re-point their imports at the shared
locations via the project's baseUrl alias.

AdvancedSwapDetails itself stayed in place because its body pulls
two view-local siblings (./FormattedPriceImpact, ./styleds) that
also duplicate in both views. Pulling the styled chain into
components/ is a larger move worth its own follow-up.
