+++ src/utils/confirmPriceImpactWithoutFee.ts
@@ -1,5 +1,5 @@
 import { Percent } from '@pancakeswap/sdk'
-import { ALLOWED_PRICE_IMPACT_HIGH, PRICE_IMPACT_WITHOUT_FEE_CONFIRM_MIN } from '../../../../config/constants'
+import { ALLOWED_PRICE_IMPACT_HIGH, PRICE_IMPACT_WITHOUT_FEE_CONFIRM_MIN } from '../config/constants'
 
 /**
  * Given the price impact, get user confirmation.
