+++ src/views/Foundation/Donate/index.tsx
@@ -11,10 +11,10 @@ import ConfirmSwapModal from './components/ConfirmSwapModal'
 import CurrencyInputPanel from '../../../components/CurrencyInputPanel'
 import { AutoRow, RowBetween } from '../../../components/Layout/Row'
 import AdvancedSwapDetailsDropdown from './components/AdvancedSwapDetailsDropdown'
-import confirmPriceImpactWithoutFee from './components/confirmPriceImpactWithoutFee'
+import confirmPriceImpactWithoutFee from 'utils/confirmPriceImpactWithoutFee'
 import { SwapCallbackError, Wrapper } from './components/styleds'
 import TradePrice from './components/TradePrice'
-import ProgressSteps from './components/ProgressSteps'
+import ProgressSteps from 'components/ProgressSteps'
 import { AppHeader, AppBody } from '../../../components/App'
 import ConnectWalletButton from '../../../components/ConnectWalletButton'
 import useActiveWeb3React from '../../../hooks/useActiveWeb3React'
